Relay: parse crc32 as unsigned (fw 0.24.8)
A fully transferred HiveInside image was rejected by the node with
OTA_ERR_CRC (state=0x13). Not OTA_ERR_SIZE, and the node checks size
first, so all 169540 bytes had arrived intact — the two ends simply
disagreed about the expected CRC.

The relay parsed it as `payload["crc32"] | 0`. ArduinoJson deduces the
type from the default operand, and `int` is 32-bit here, so is<int>() is
false for any value above 2147483647 and the fallback is taken. The hub
sent crc32 = 0 in BEGIN; the node computed the real CRC and refused the
image.

Roughly half of all CRC-32 values have the high bit set, so the relay has
been a coin flip since it was written: it succeeded with hiveinside 0.3.5
(crc 128852642) and fails with 0.4.2 (crc 2602123579). The hub's own
self-update path used `| 0U` and was never affected.

FIRMWARE_VERSION goes to 0.24.8 so the fixed build is actually newer than
what devices run — the OTA gate serves nothing that is not strictly newer,
so shipping this under 0.24.7 would leave every hub on the broken build
while the dashboard showed them up to date.

Tests: the server side must keep putting the CRC on the wire as an exact
integer, never truncated, clamped or stringified. The parse itself cannot
be covered here — reproducing it needs ArduinoJson's type deduction, and
the host-test harness is deliberately toolchain-free — so the hazard is
documented at the call site instead.

+++ firmware/src/hivehub_network.cpp
@@ -1069,7 +1069,15 @@ void checkCommands() {
     int slot = payload["slot"] | 1;
     String mac = hivecfg::hiveInsideMacForSlot((uint8_t)slot);
     String fwUrl = payload["url"] | "";
-    uint32_t crc = (uint32_t)(payload["crc32"] | 0);
+    // `| 0U`, never `| 0`: ArduinoJson deduces the type from the default, and a
+    // signed `int` is 32-bit here, so is<int>() is false for any CRC above
+    // 2147483647 and the fallback (0) is taken. The node was then told to expect
+    // crc32 = 0, computed the real one, and rejected a perfectly transferred
+    // image with OTA_ERR_CRC (state=0x13). Since roughly half of all CRC-32
+    // values exceed 2^31 this made the relay a coin flip — it worked with
+    // hiveinside 0.3.5 (crc 128852642) and failed with 0.4.2 (crc 2602123579).
+    // The hub's own self-update path had this right (`| 0U`) all along.
+    uint32_t crc = payload["crc32"] | 0U;
     if (fwUrl.length() == 0) {
       postCommandResult(commandId, false, "update_hiveinside missing url");
     } else if (mac.length() == 0) {
